requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
greenstalk==2.0.2
openai-whisper==20231117
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2
python-dotenv==1.0.0
alembic==1.13.0
pymysql==1.1.0
